Refactor code in order to resolve code climate issues
